stats_service/requirements.txt
aiokafka
clickhouse-driver
grpcio
grpcio-tools
protobuf
pydantic_settings